pyproject.toml
[project]
name = "project01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "exa-py>=2.11.0",
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.4",
]
